Fixed bug where replay would out of sync on first command on linux
Linux C++ appears to be truncating overflows in a strange way.
LuatoNumber is signed, but SyncSeed is unsigned, so when you import the number
it comes in incorrectly.  It's strictly typed 1 way an then the other to ensure
there are no problems.

I believe GameCycle also has the problem, but we don't usually have games
long than 2 billion cycles.