Show terrain flags in status line of map editor
In the map editor, if the user points the mouse cursor at a patch
in the map, the editor displays the type and position of the patch
in the status line.  Make it also describe the flags of the map
field at which the user is pointing.

languages/genpot.sh now uses --add-comments and presumably requires
GNU gettext.  I don't know whether it worked with others before.