Preserve WEAK_TARGET flag in more places
This was a bug in engine/action/action_attack.cpp.  The priority/cost
computations did not get called because the attacking unit forgot that
its goal was weak.  One way to trigger the bug was by telling an
assault unit to go attack a place near an enemy camp.  The assault
unit moved part of the way, then noticed an enemy engineer near the
camp, took him as a weak goal, and killed him.  Next, SubAction became
0 again, and the assault unit was about to continue to the ordered
place, but HandleActionAttack noted that SubAction was 0, and called
CheckForTargetInRange, which found the enemy training camp and set
that as another weak goal, setting unit->SubAction |= WEAK_TARGET.
SubAction thus became 2 = WEAK_TARGET, but HandleActionAttack (still
in "case 0:") noticed that the assault unit now had a goal (weak or
not), and because the training camp was close enough, began also
attacking it and set unit->SubAction = ATTACK_TARGET.  And that's
where the WEAK_TARGET flag was lost.  I guess the same could have
happened without the enemy engineer.

Fix by preserving the WEAK_TARGET flag in each place where SubAction
is changed, unless the flag must be clear there (in which case, assert
that), or the whole order is being replaced (from SavedOrder or just
cleared out).

Fixes: http://savannah.nongnu.org/bugs/?30971
 (Units attack buildings too stubbornly)