Move common code to new func TriggerMatchUnitType.
This changes the behavior of CclGetNumUnitsAt a little.  It now checks
unit->Constructed even if unittype is a category rather than an
individual type.  This does not affect the bundled maps because the
only use of GetNumUnitsAt is in campaigns/swindler/level01.sms,
which searches for "unit-assault", and assault units do not appear
on the map before they are fully trained.