Replaced AiList and _ai_scripts_ with AiTypes
Also replaced RegisterAi and DefineAi with DefineAiType,
which now takes a table rather than a set of parameters.
This table is the object that DefineAiType adds to AiTypes.
The keys of the table are constant strings, not numbers.

The purpose is to make the engine more aware of AI initialization
functions, although this commit doesn't yet get that far.